requirements.txt
requests
pytest
pytest-HTML
mysql-connector-python
